configs: TIPO MoE-1B production run is fp16 + mxfp8 + Muon
GRAD_SCALER="auto" turns loss scaling on whenever fp16 appears in the parameter
or autocast dtype, so bf16 configs pay nothing and need no branch.

Measured over 24 steps on 4x5090: loss 3.26 at step 24 against the bf16
baseline's 3.15, load imbalance falling 7.81 -> 5.74, scale settled at 32768
with no overflows after the initial backoff.

+++ configs/lm/tipo_moe_1b_uwupipe.py
@@ -55,10 +55,12 @@ NUM_MICROBATCHES = 16
 # [5,4,4,3] measured 13% slower here than the [5,5,5,1] measurement derives.
 LAYERS = []
 SCHEDULE = "1f1b"
-PARAM_DTYPE = "bf16"
-AUTOCAST_DTYPE = "bf16"
+PARAM_DTYPE = "fp16"
+AUTOCAST_DTYPE = "fp16"
 MXFP8 = True
 
+# Muon on the hidden matrices; its non-matrix group is AdamW over the 16-bit
+# kernel, which is what makes it fp16-safe. See docs/internals/optimizers.md.
 OPTIMIZER = "muon"
 OPTIMIZER_KWARGS = {"muon_lr": 2e-3, "embed_lr": 2e-3}
 LR = 5e-4
